+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.0.13 / Unreleased
+
+[FEATURE] Implement Searcher#records
+[FEATURE] Implement Searcher#ids
+[FEATURE] Allow for Array-style method iteration
+[FEATURE] Changed syntax to match ActiveRecord style.
+
 0.0.12 / 2013-10-24
 
 [FEATURE] Allow filter values to be arrays.